# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Register API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1140",
    "paths": {
...
        "/user/edit-profile": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.VolunteerPersonalData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.SuccessResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_Orken1119_Register_internal_models.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.Properties"
                }
            }
        },
        "github_com_Orken1119_Register_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.ErrorDetail"
                    }
                }
            }
        },
...
        "github_com_Orken1119_Register_internal_models.Properties": {
            "type": "object",
            "properties": {
                "additionalProp1": {
                    "type": "string"
                },
                "additionalProp2": {
                    "type": "string"
                },
                "additionalProp3": {
                    "type": "string"
                }
            }
        },
        "github_com_Orken1119_Register_internal_models.SuccessResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/github_com_Orken1119_Register_internal_models.Properties"
                },
                "result": {}
            }
        },
...
        "github_com_Orken1119_Register_internal_models.VolunteerPersonalData": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "skills": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}